+++ cli/commands/asset/delete_test.go
@@ -1,0 +1,80 @@
+package asset
+
+import (
+	"errors"
+	"testing"
+
+	client "github.com/sensu/sensu-go/cli/client/testing"
+	test "github.com/sensu/sensu-go/cli/commands/testing"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+)
+
+func TestDeleteCommand(t *testing.T) {
+	assert := assert.New(t)
+
+	cli := test.NewMockCLI()
+	cmd := DeleteCommand(cli)
+	require.NoError(t, cmd.Flags().Set("skip-confirm", "t"))
+
+	assert.NotNil(cmd, "cmd should be returned")
+	assert.NotNil(cmd.RunE, "cmd should be able to be executed")
+	assert.Regexp("delete", cmd.Use)
+	assert.Regexp("asset", cmd.Short)
+}
+
+func TestDeleteCommandRunEClosureWithoutName(t *testing.T) {
+	assert := assert.New(t)
+
+	cli := test.NewMockCLI()
+	cmd := DeleteCommand(cli)
+	require.Error(t, cmd.Flags().Set("timeout", "15"))
+	require.NoError(t, cmd.Flags().Set("skip-confirm", "t"))
+	out, err := test.RunCmd(cmd, []string{})
+
+	assert.Regexp("Usage", out) // usage should print out
+	assert.Error(err)
+}
+
+func TestDeleteCommandRunEClosureWithFlags(t *testing.T) {
+	assert := assert.New(t)
+
+	cli := test.NewMockCLI()
+	client := cli.Client.(*client.MockClient)
+	client.On("Delete", mock.Anything, mock.Anything).Return(nil)
+
+	cmd := DeleteCommand(cli)
+	require.NoError(t, cmd.Flags().Set("skip-confirm", "t"))
+	out, err := test.RunCmd(cmd, []string{"my-asset"})
+
+	assert.Regexp("Deleted", out)
+	assert.Nil(err)
+}
+
+func TestDeleteCommandRunEClosureWithServerErr(t *testing.T) {
+	assert := assert.New(t)
+
+	cli := test.NewMockCLI()
+	client := cli.Client.(*client.MockClient)
+	client.On("Delete", mock.Anything, mock.Anything).Return(errors.New("oh noes"))
+
+	cmd := DeleteCommand(cli)
+	require.NoError(t, cmd.Flags().Set("skip-confirm", "t"))
+	out, err := test.RunCmd(cmd, []string{"test-handler"})
+
+	assert.Empty(out)
+	assert.NotNil(err)
+	assert.Equal("oh noes", err.Error())
+}
+
+func TestDeleteCommandRunEFailConfirm(t *testing.T) {
+	assert := assert.New(t)
+
+	cli := test.NewMockCLI()
+	cmd := DeleteCommand(cli)
+	out, err := test.RunCmd(cmd, []string{"test-handler"})
+	require.NoError(t, err)
+
+	assert.Contains(out, "Canceled")
+}